convert drag service to a function
