+++ scripts/apps/archive/services/index.ts
@@ -1,4 +1,3 @@
-export {DragItemService} from './DragItemService';
 export {FamilyService} from './FamilyService';
 export {MultiService} from './MultiService';
 export {ArchiveService} from './ArchiveService';
